Retail Marketing exploratory Data Analysis and Data Preprocessing
The project is based on exploratory data analysis and data preprocessing methods to understand the marketing campaigns and their outcomes. It involves making use of univariate and bivariate analysis, visualization methods, garbage value treatment, and data manipulation methods to answer queries related to cosumer's purchasing pattern and prepare a report for the management team.The project is based on exploratory data analysis and data preprocessing methods to understand the marketing campaigns and their outcomes. It involves making use of univariate and bivariate analysis, visualization methods, garbage value treatment, and data manipulation methods to answer queries related to cosumer's purchasing pattern and prepare a report for the management team.
Skills: Data Visualization · Data Cleaning